requirements.txt
langchain
openai
langchain_openai
# dotenv
telebot
langchainhub
openai-whisper
faiss-cpu
transformers